requirements.txt
opencv-python
numpy
scikit-learn
matplotlib
PyYAML